lesson7 added

+++ market/src/main/java/ru/geekbrains/market/repositories/specifications/ProductSpecifications.java
@@ -21,10 +21,10 @@ public class ProductSpecifications {
     public static Specification<Product> build(MultiValueMap<String, String> params) {
         Specification<Product> spec = Specification.where(null);
         if (params.containsKey("min_price") && !params.getFirst("min_price").isBlank()) {
-            spec = spec.and(ProductSpecifications.priceGreaterOrEqualsThan(Integer.parseInt(params.getFirst("min_price"))));
+            spec = spec.and(ProductSpecifications.priceGreaterOrEqualsThan(Double.parseDouble(params.getFirst("min_price"))));
         }
         if (params.containsKey("max_price") && !params.getFirst("max_price").isBlank()) {
-            spec = spec.and(ProductSpecifications.priceLesserOrEqualsThan(Integer.parseInt(params.getFirst("max_price"))));
+            spec = spec.and(ProductSpecifications.priceLesserOrEqualsThan(Double.parseDouble(params.getFirst("max_price"))));
         }
         if (params.containsKey("title") && !params.getFirst("title").isBlank()) {
             spec = spec.and(ProductSpecifications.titleLike(params.getFirst("title")));
